remove routes and return to deployment
