--- .env
@@ -1,1 +1,0 @@
-JWT_SECRET_KEY="superSecretKey"
